added user privileges to users.html

+++ src/main/java/org/jasr/dashard/controller/UserController.java
@@ -1,6 +1,7 @@
 package org.jasr.dashard.controller;
 
 import java.io.IOException;
+import java.security.Principal;
 
 import org.jasr.dashard.dao.UserDAO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,14 +17,16 @@ public class UserController {
     private UserDAO userDAO;
 
     @RequestMapping(value = "/upsert", method = RequestMethod.POST)
-    public String upsert(String username, String password) throws IOException {
-        userDAO.upsert(username, password);
+    public String upsert(Principal principal, String username, String password) throws IOException {
+        if ("admin".equals(principal.getName()))
+            userDAO.upsert(username, password);
         return "redirect:/users.html";
     }
 
     @RequestMapping(value = "/delete", method = RequestMethod.GET)
-    public String delete(String username) throws IOException {
-        userDAO.delete(username);
+    public String delete(Principal principal, String username) throws IOException {
+        if ("admin".equals(principal.getName()))
+            userDAO.delete(username);
         return "redirect:/users.html";
     }
 }
